+++ src/dlc/UI/Shared/CivVAccess_BaseMenuHelp.lua
@@ -21,6 +21,10 @@ BaseMenuHelp.ListNavHelpEntries = {
         description = "TXT_KEY_CIVVACCESS_HELP_DESC_REVIEW_SECTIONS",
     },
     {
+        keyLabel = "TXT_KEY_CIVVACCESS_HELP_KEY_ALT_HOME_END",
+        description = "TXT_KEY_CIVVACCESS_HELP_DESC_REVIEW_FIRST_LAST",
+    },
+    {
         keyLabel = "TXT_KEY_CIVVACCESS_HELP_KEY_HOME_END",
         description = "TXT_KEY_CIVVACCESS_HELP_DESC_JUMP_FIRST_LAST",
     },